policykit: Update to 0.120 + 3 additional commits
Add rudimentary support for the meson build system to the
gobject_introspection PortGroup, which is used by polkit. This may need
further adaption in case other build systems do not name their gobject
introspection configure flag 'introspection'.

Switch polkit to the meson build system, drop the autoconf related files
and sections from the Portfile. Re-enable the livecheck that seems to
have been disabled because of warning in upstream's NEWS file that has
been also present in previous versions and was not a good reason to keep
polkit at an older version.

Fix CVE-2021-4034, a local privilege escalation in pkexec – although
I have not verified whether this is exploitable on macOS.

The updated polkit would require a newer mozjs – however, polkit
upstream is working on replacing polkit with the much smaller and
simpler to build duktape JavaScript engine, which will ship in 0.121.
Instead of packaging the newer version of mozjs, I packaged duktape and
pulled in the commit the enables using it. See
  https://gitlab.freedesktop.org/polkit/polkit/-/merge_requests/97
for further discussion that suggests that pretty much all distributions
will switch to duktape as soon as it is available.

CVE: CVE-2021-4034
Closes: https://trac.macports.org/ticket/63277